Add `aria_hidden` option
Closes [#72].

Sometimes SVGs are simply decorative and therefore should be removed
from the accessibility tree. Here’s an example:

```erb
<p>
  <%= inline_svg "checkmark-icon.svg" %>
  Success!
</p>
```

You can imagine this as a simple flash message, letting the user know
that their action succeeded. We have the word ‘Success’, preceded by a
checkmark icon to visually represent the state.

In such a situation, the icon is simply decorative; it’s repeating
content already conveyed by the word ‘Success.’ If we apply
`aria-hidden="true"` to the SVG, we can remove it from the accessibility
tree and avoid the repetition.

[#72]: https://github.com/jamesmartin/inline_svg/issues/72