Set asset finder from callable
It fixes #151 by allowing the asset finder to be a callable object. This is
useful when the asset pipeline is not ready when the initializer is run, for
example, when using Propshaft.

I don't know if this is the best way to solve this problem. On the one hand,
Propshaft could be done when its initializer is run, but on the other hand,
this gem's railtie assumes the asset pipeline is ready.